requirements.txt
Django==5.1.3
djangorestframework==3.15.2
psycopg2-binary==2.8.6
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.2.0